(#7188) Rearrange dump mounttab parameter to facilitate testing
Trying to stub Facter when it's used directly in the newproperty block
doesn't work, since the block is evaluated before we're able to stub
it in the rspec tests.

Because of this, we "allow" all possible valid values, and check for
the operating-system specific restrictions in the validate block,
since it isn't evaluated until run-time.

Paired-with: Max Martin <max@puppetlabs.com>