(#13070) Add module metadata
This patch adds the Modulefile and README.markdown.  Without this patch,
this repository does not follow the conventions or the scaffolding of
`puppet module generate` which is a problem because this module is
inconsistent with other modules.

This patch fixes the problem by adding module metadata suitable for
`puppet module build`