Add puppetlabs_spec_helper support
Without this patch the spec tests are directly coupled to Puppet and
Facter.  This is a problem because it makes it difficult to test the
module against a configuration matrix containing various Puppet and
Facter releases.

This patch fixes the problem by delegating the initialization of Puppet
and Facter to the puppetlabs_spec_helper [1] repository.

[1] https://github.com/puppetlabs/puppetlabs_spec_helper