(maint) Add default mountpoint provider
Without this patch the types and providers for the mountpoint do not
specify the default systems they're useful for.  This patch is a
maintenance change only that cleans up the layout of the mountpoint
providers.

This patch also fixes the existing spec tests for the linux and solaris
providers to prevent order dependant failures.  Without this patch
applied the resources being tested do not explicitly set the provider
that should be used.  This patch fixes the problem by passing the
`:provider => :solaris` and `:provider => :linux` arguments directly to
the Type initializer.