(maint) Fix bad spec test for Linux provider
Without this patch the mountpoint type linux provider is failing the
spec test.  The reason this is failing is because the test is
incorrectly specifying that the resource should exist on the system when
the example provided means it should not.

The test reads "should be present if it is included in mount output with
an incorrect device" (keyword device) but the mock mount output changes
the mount directory and not the device.

This patch fixes the problem by keeping the mount directory fixed and
providing an incorrect device.

The failing test that should pass with this patch applied is:

    Failures:

      1) Puppet::Type::Mountpoint::ProviderLinux#exists? should be present if it is included in mount output with an incorrect device
         Failure/Error: provider.should be_exists
           expected exists? to return true, got false
         # ./spec/unit/provider/mountpoint/linux_spec.rb:62