(Maint) Fix up spec tests to run properly
Without this patch the spec tests for the mount providers are not
running with recent versions of Puppet.  This is because this module
uses and old version of the puppet spec helper.

This patch fixes the problem by updating the spec tests to use the new
spec helper gem library.