(#3530) Fix mountpoint; make it refreshable
Without this patch the mountpoint type and provider is not capable of
reacting to refresh messages sent by subscribe or notify events.  This
is a problem because end users expect to be able to notify the
mountpoint resource in order to remount the filesystem modeled by the
resource.

We determined the cause of this problem was that the refresh method was
implemented in the provider classes instead of the type.

This patch fixes the problem by implementing the refresh method for the
mountpoint type.

We also deleted the spec tests that appeared to exercise the refresh
behavior for solaris.  After reviewing the Puppet codebase, we found no
possible way the refresh method would be called on the provider.

Paired-with: Henrik Lindberg <henrik.lindberg@cloudsmith.com>